+++ src/renderer/components/ExamScreen/Explanation.js
@@ -1,16 +1,13 @@
 import React from 'react'
 import { shell } from 'electron'
-import { withStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
 import Button from '@material-ui/core/Button'
 import OpenIcon from '@material-ui/icons/OpenInNewSharp'
 import decodeCorrect from '../../utils/decodeCorrect'
 import isEqual from 'lodash/isEqual'
 
-const styles = theme => ({})
-
-function Explanation({ expRef, variant, answers, correctAnswers, explanation, classes }) {
-  const status = variant === 2 ? answers[0] : isEqual(answers, correctAnswers)
+function Explanation({ expRef, variant, answers, correctAnswers, explanation }) {
+  let status = variant === 0 || variant === 1 ? isEqual(answers, correctAnswers) : answers[0]
   return (
     <div ref={expRef} className="explanation">
       <Typography variant="subtitle1">
@@ -27,23 +24,23 @@ function Explanation({ expRef, variant, answers, correctAnswers, explanation, cl
         {explanation.length > 0 &&
           explanation.map((e, i) => {
             if (e.variant === 0) {
-              return <img key={`explanation - ${i}`} src={e.text} alt="" />
+              return <img key={i} src={e.text} alt="" />
             } else if (e.variant === 1) {
               return (
-                <Typography key={`explanation - ${i}`} variant="subtitle1">
+                <Typography key={i} variant="subtitle1">
                   {e.text}
                 </Typography>
               )
             } else if (e.variant === 2) {
               return (
-                <Typography key={`explanation - ${i}`} variant="h5">
+                <Typography key={i} variant="h5">
                   {e.text}
                 </Typography>
               )
             } else if (e.variant === 3) {
               return (
                 <Button
-                  key={`explanation - ${i}`}
+                  key={i}
                   className="explanation-link"
                   onClick={() => shell.openExternal(e.href)}
                 >
@@ -58,4 +55,4 @@ function Explanation({ expRef, variant, answers, correctAnswers, explanation, cl
   )
 }
 
-export default withStyles(styles)(Explanation)
+export default Explanation